Refactor CGaz
