+++ src/cgame/cg_local.h
@@ -2383,13 +2383,13 @@ extern vmCvar_t cg_nofatigue;
 extern vmCvar_t com_maxfps;
 extern vmCvar_t com_hunkmegs;
 
-extern vmCvar_t cg_drawCGaz;
-extern vmCvar_t cg_CGazY;
-extern vmCvar_t cg_CGazHeight;
-extern vmCvar_t cg_CGazWidth;
+extern vmCvar_t etj_drawCGaz;
+extern vmCvar_t etj_CGazY;
+extern vmCvar_t etj_CGazHeight;
+extern vmCvar_t etj_CGazWidth;
 extern vmCvar_t etj_CGazColor1;
 extern vmCvar_t etj_CGazColor2;
-extern vmCvar_t cg_CGazAlpha;
+extern vmCvar_t etj_CGazAlpha;
 extern vmCvar_t cg_drawCGazUsers;
 
 extern vmCvar_t cg_drawOB;
@@ -3966,6 +3966,7 @@ namespace ETJump
 	int checkExtraTrace(int value);
 	void onPlayerRespawn(qboolean revived);
 	void runFrameEnd();
+	void DrawCGazHUD();
 
 	enum extraTraceOptions {
 		OB_DETECTOR,
